+++ bfabric/scripts/bfabric_list_not_available_proteomics_workunits.py
@@ -20,11 +20,12 @@ from loguru import logger
 from rich.console import Console
 from rich.table import Column, Table
 
-from bfabric import Bfabric, BfabricClientConfig
+from bfabric import Bfabric
 from bfabric.cli_formatting import setup_script_logging
+from bfabric.entities import Parameter
 
 
-def render_output(workunits_by_status: dict[str, list[dict[str, Any]]], config: BfabricClientConfig) -> None:
+def render_output(workunits_by_status: dict[str, list[dict[str, Any]]], client: Bfabric) -> None:
     """Renders the output as a table."""
     table = Table(
         Column("AID", no_wrap=True),
@@ -33,7 +34,12 @@ def render_output(workunits_by_status: dict[str, list[dict[str, Any]]], config: 
         Column("Status", no_wrap=True),
         Column("Created by", no_wrap=True, max_width=12),
         Column("Name", no_wrap=False),
+        Column("Nodelist", no_wrap=False),
     )
+
+    workunit_ids = [wu["id"] for wu_list in workunits_by_status.values() for wu in wu_list]
+    nodelist_params = Parameter.find_by({"workunitid": workunit_ids, "key": "nodelist"}, client)
+    nodelist_values = {param["workunit"]["id"]: param.value for param in nodelist_params.values()}
 
     for status, workunits_all in workunits_by_status.items():
         workunits = [x for x in workunits_all if x["createdby"] not in ["gfeeder", "itfeeder"]]
@@ -44,8 +50,9 @@ def render_output(workunits_by_status: dict[str, list[dict[str, Any]]], config: 
         }.get(status, "black")
 
         for wu in workunits:
-            app_url = f"{config.base_url}/application/show.html?id={wu['application']['id']}"
-            wu_url = f"{config.base_url}/workunit/show.html?id={wu['id']}&tab=details"
+            app_url = f"{client.config.base_url}/application/show.html?id={wu['application']['id']}"
+            wu_url = f"{client.config.base_url}/workunit/show.html?id={wu['id']}&tab=details"
+
             table.add_row(
                 f"[link={app_url}]A{wu['application']['id']:3}[/link]",
                 f"[link={wu_url}]WU{wu['id']}[/link]",
@@ -53,6 +60,7 @@ def render_output(workunits_by_status: dict[str, list[dict[str, Any]]], config: 
                 f"[{status_color}]{status}[/{status_color}]",
                 wu["createdby"],
                 wu["name"],
+                nodelist_values.get(wu["id"], "N/A"),
             )
 
     console = Console()
@@ -76,7 +84,7 @@ def list_not_available_proteomics_workunits(date_cutoff: datetime) -> None:
             obj={"status": status, "createdafter": date_cutoff.isoformat()},
         ).to_list_dict()
 
-    render_output(workunits_by_status, config=client.config)
+    render_output(workunits_by_status, client=client)
 
 
 def main() -> None: