+++ docs/changelog.md
@@ -13,8 +13,11 @@ Versioning currently follows `X.Y.Z` where
 ### Added
 
 - The `Bfabric` instance is now pickleable.
-- More entities, and the ability to define circular relationships by deferring imports.
-- `Entity.id` is now a property. It is still not intended to add all properties, or defer them all.
+- Entities mapping:
+  - More types and relationships
+  - Relationships defer imports to descriptor call, i.e. circular relationships are possible now.
+  - Add `Entity.id` and `Entity.web_url` properties.
+  - `HasOne` allows defining `optional=True`.
 
 ## \[1.13.4\] - 2024-08-05