+++ docs/changelog.md
@@ -14,9 +14,10 @@ Versioning currently follows `X.Y.Z` where
 
 - The `Bfabric` instance is now pickleable.
 - Entities mapping:
+  - Add `Entity.id` and `Entity.web_url` properties.
+  - Add `Entity.__getitem__` to access fields from the data dictionary directly.
   - More types and relationships
   - Relationships defer imports to descriptor call, i.e. circular relationships are possible now.
-  - Add `Entity.id` and `Entity.web_url` properties.
   - `HasOne` and `HasMany` allow defining `optional=True` to indicate fields which can be missing under some circumstances.
 
 ## \[1.13.4\] - 2024-08-05