+++ docs/changelog.md
@@ -8,7 +8,7 @@ Versioning currently follows `X.Y.Z` where
 - `Y` should be the current bfabric release
 - `Z` is increased for feature releases, that should not break the API
 
-## \[tba\] - tba
+## \[Unreleased\]
 
 ### Added